+++ saleor/graphql/order/mutations/orders.py
@@ -259,7 +259,11 @@ class OrderUpdateShipping(BaseMutation):
         clean_order_update_shipping(order, method)
 
         order.shipping_method = method
-        order.shipping_price = info.context.plugins.calculate_order_shipping(order)
+        shipping_price = info.context.plugins.calculate_order_shipping(order)
+        order.shipping_price = shipping_price
+        order.shipping_tax_rate = info.context.plugins.get_order_shipping_tax_rate(
+            order, shipping_price
+        )
         order.shipping_method_name = method.name
         order.save(
             update_fields=[
@@ -268,6 +272,7 @@ class OrderUpdateShipping(BaseMutation):
                 "shipping_method_name",
                 "shipping_price_net_amount",
                 "shipping_price_gross_amount",
+                "shipping_tax_rate",
             ]
         )
         update_order_prices(order, info.context.discounts)