+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Drop authorization keys API - #6631 by @maarcingebala
 - Add metadata to menu and menu item - #6648 by @tomaszszymanski129
 - Get tax rate from plugins - #6649 by @IKarbowiak
+- Add order shipping tax rate - #6678 by @IKarbowiak
 
 # 2.11.1